fishnet_cod/requirements.txt
aars~=0.3.3
aiohttp
aleph-sdk-python==0.6.0
fastapi
pandas
pydantic
semver